helm chart - added initContainer to post-job
As per kubernetes best practice, a job can be held off until the resources it depends on become available. This is cleaner than a time out or a try-loop in the actual job: the currently implemented jobs wait for 30 seconds and then crash. On slower systems, the post-job can fail. With this initContainer the post-job always waits long enough and does not fail.